% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{LdFlags}
\alias{LdFlags}
\title{Output the LD flags for linking against pb. 
It is only needed to be called for windows.
For nix-OS, the other package simply import this for the side effect of loading RProtoBuflib.}
\usage{
LdFlags()
}
\description{
Output the LD flags for linking against pb. 
It is only needed to be called for windows.
For nix-OS, the other package simply import this for the side effect of loading RProtoBuflib.
}
